+++ packages/patternfly-4/react-core/src/components/Avatar/Avatar.test.tsx
@@ -3,6 +3,6 @@ import { shallow } from 'enzyme';
 import Avatar from './Avatar';
 
 test('simple avatar', () => {
-  const view = shallow(<Avatar alt="avatar" />);
+  const view: any = shallow(<Avatar alt="avatar" />);
   expect(view).toMatchSnapshot();
 });