+++ packages/patternfly-4/react-core/src/components/index.ts
@@ -18,7 +18,7 @@ export * from './List';
 export * from './LoginPage';
 export * from './Modal';
 export * from './Nav';
-// export * from './Popover'; // Not ready yet
+export * from './Popover';
 export * from './Progress';
 export * from './Radio';
 export * from './Select';