+++ packages/patternfly-4/react-core/typings/react-styles.d.ts
@@ -1,0 +1,4 @@
+declare module '*.css' {
+  const styles: any;
+  export default styles;
+}