+++ packages/patternfly-4/react-docs/gatsby-config.js
@@ -5,6 +5,7 @@ module.exports = {
     title: 'PatternFly React'
   },
   plugins: [
+    'gatsby-plugin-typescript',
     'gatsby-plugin-react-helmet',
     {
       resolve: `gatsby-source-filesystem`,