+++ packages/patternfly-4/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.29.1"></a>
+## [1.29.1](https://github.com/patternfly/patternfly-react/compare/patternfly4-react-lerna-root@1.29.0...patternfly4-react-lerna-root@1.29.1) (2018-12-10)
+
+
+
+
+**Note:** Version bump only for package patternfly4-react-lerna-root
+
 <a name="1.29.0"></a>
 # [1.29.0](https://github.com/patternfly/patternfly-react/compare/patternfly4-react-lerna-root@1.28.6...patternfly4-react-lerna-root@1.29.0) (2018-12-07)